Add build system and headless test
